+++ seacurity/src/main/java/xxk/wuhai/seacurity/work/bean/ApprovalRecordListBean.java
@@ -29,7 +29,7 @@ public class ApprovalRecordListBean {
      */
 
     private String apName;
-    private int apUserId;
+    private long apUserId;
     private int approvalRecordId;
     private String createTime;
     private String lrBeginTime;
@@ -40,7 +40,7 @@ public class ApprovalRecordListBean {
     private String status;
     private String supplement;
     private String typeId;
-    private int userId;
+    private long userId;
     private List<String> pictureUrls;
 
     public String getApName() {
@@ -51,11 +51,11 @@ public class ApprovalRecordListBean {
         this.apName = apName;
     }
 
-    public int getApUserId() {
+    public long getApUserId() {
         return apUserId;
     }
 
-    public void setApUserId(int apUserId) {
+    public void setApUserId(long apUserId) {
         this.apUserId = apUserId;
     }
 
@@ -139,11 +139,11 @@ public class ApprovalRecordListBean {
         this.typeId = typeId;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
